Testing Cart

+++ src/controllers/cart.controllers.js
@@ -2,33 +2,51 @@ const catchError = require('../utils/catchError');
 const Cart = require('../models/Cart');
 const User = require('../models/User');
 const Product = require('../models/Product');
+const Category = require('../models/Category');
 
 const getAll = catchError(async(req, res) => {
-    const { id } = req.params
+    const userId = req.user.id    
     const results = await Cart.findAll({
-        include:[User, Product],
-        where: {userId, productId}
+        where: {userId},
+        include:[
+            {
+                model: Product,
+                attributes: {exclude:['createdAt', ]},
+                include:{
+                    model: Category,
+                    attributes:['name']
+
+                }
+            }
+        ]
     });
     return res.json(results);
 });
 
 const create = catchError(async(req, res) => {
-    const result = await Cart.create(req.body);
+    const userId = req.user.id
+    const { quantity, productId } = req.body
+
+    const newBody = { userId, quantity, productId }
+    const result = await Cart.create(newBody);
     return res.status(201).json(result);
 });
 
 const remove = catchError(async(req, res) => {
     const { id } = req.params;
-    const result = await Cart.destroy({ where: {id} });
+    const userId = req.user.id
+    const result = await Cart.destroy({ where: {id, userId} });
     if(!result) return res.sendStatus(404);
     return res.sendStatus(204);
 });
 
 const update = catchError(async(req, res) => {
+    const userId = req.user.id
+    const { quantity } = req.body
     const { id } = req.params;
     const result = await Cart.update(
-        req.body,
-        { where: {id}, returning: true }
+        { quantity },
+        { where: {id, userId}, returning: true }
     );
     if(result[0] === 0) return res.sendStatus(404);
     return res.json(result[1][0]);
